Correct Perl errors and English errors
1.  Perl errors.  The script attempted to remove the last component from the path of each application on which it depends, such as in the line: $repeatmasker=~s/RepeatMasker\n?// unless -d $repeatmasker;
However, this command removes the FIRST occurrence of "RepeatMasker" in the string, not the last.  This was causing LTR_retriever to fail in our installation (where the path to RepeatMasker is: /home/apps/eb/software/RepeatMasker/4.1.2-p1-foss-2020b/RepeatMasker).  I have therefore "anchored" the search expression to the end of the string by adding a "$" to each such expression.  

(I do feel the logic here is cumbersome: the script removes the last component from the path of each application, only to add it back later.)  

2.  English errors.  There were a few occurrences of the phrase "is not exist" in error messages, which is not correct English.  I have amended these to "does not exist".